requirements.txt
Django==3.2.16
python-dotenv==0.21.0
stripe==5.0.0
